gpt ver 5 some code doesnt work
""In this updated code, a fourth tab called Tab 4 has been added. The background color of this tab is set to rgba(77, 219, 255, 1) by creating a palette and applying it to the tab.

Inside Tab 4, a button called exit_button has been added. Clicking this button will call the exit_program() method, which terminates the program using sys.exit().

The exit_program() method is defined to handle the termination of the program.

Additionally, a few additional methods (changeEvent() and closeEvent()) have been added to ensure the window receives focus when activated and to override the close event behavior so that the program is terminated only when the exit button is clicked.""